% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{archspirals}
\alias{archspirals}
\title{Persistence diagrams from Archimedean spiral samples}
\format{
An object of class \code{persistence_set} containing 24 objects of class
\code{\link[phutil:persistence]{phutil::persistence}}.
}
\usage{
archspirals
}
\description{
A set of 24 persistence diagrams computed from noisy samples of
2-armed Archimedean spirals. Each sample consists of 120 points sampled
from an Archimedean spiral, embedded in 3D with a zero z-coordinate, then
Gaussian noise (sd = 0.05) added. Vietoris-Rips persistence was computed up
to dimension 2 with maximum scale 6 using \code{\link[TDA:ripsDiag]{TDA::ripsDiag()}}. Generated
with seed \code{28415}.
}
\keyword{datasets}
