% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inphr-package.R
\docType{package}
\name{inphr-package}
\alias{inphr}
\alias{inphr-package}
\title{inphr: Statistical Inference for Persistence Homology Data}
\description{
A set of functions for performing null hypothesis testing on samples of persistence diagrams using the theory of permutations. Currently, only two-sample testing is implemented. Inputs can be either samples of persistence diagrams themselves or vectorizations. In the former case, they are embedded in a metric space using either the Bottleneck or Wasserstein distance. In the former case, persistence data becomes functional data and inference is performed using tools available in the 'fdatest' package. Main reference for the interval-wise testing method: Pini A., Vantini S. (2017) "Interval-wise testing for functional data" \doi{10.1080/10485252.2017.1306627}. Main reference for inference on populations of networks: Lovato, I., Pini, A., Stamm, A., & Vantini, S. (2020) "Model-free two-sample test for network-valued data" \doi{10.1016/j.csda.2019.106896}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tdaverse/inphr}
  \item \url{https://tdaverse.github.io/inphr/}
  \item Report bugs at \url{https://github.com/tdaverse/inphr/issues}
}

}
\author{
\strong{Maintainer}: Aymeric Stamm \email{aymeric.stamm@cnrs.fr} (\href{https://orcid.org/0000-0002-8725-3654}{ORCID})

}
\keyword{internal}
