% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-sample-test.R
\name{two_sample_test}
\alias{two_sample_test}
\title{Two-Sample Test for Persistence Homology Data}
\usage{
two_sample_test(
  x,
  y,
  dimension = 0L,
  p = 2L,
  ncores = 1L,
  B = 1000L,
  npc = "tippett",
  verbose = FALSE
)
}
\arguments{
\item{x}{An object of class \code{persistence_set} typically produced by
\code{\link[phutil:persistence-set]{phutil::as_persistence_set()}} or of class \code{dist} typically produced by
\code{\link[phutil:pairwise-distances]{phutil::bottleneck_pairwise_distances()}} or
\code{\link[phutil:pairwise-distances]{phutil::wasserstein_pairwise_distances()}}. If \code{x} is a persistence set,
then \code{y} must be either a vector of two integers (sample sizes) or another
persistence set. If \code{x} is a distance matrix, then \code{y} must be a vector of
two integers (sample sizes).}

\item{y}{An object of class \code{persistence_set} typically produced by
\code{\link[phutil:persistence-set]{phutil::as_persistence_set()}} or a vector of two integers. If \code{x} is a
persistence set, then \code{y} must be either a vector of two integers (sample
sizes) or another persistence set. If \code{x} is a distance matrix, then \code{y}
must be a vector of two integers (sample sizes).}

\item{dimension}{An integer value specifying the homology dimension to use.
Defaults to \code{0L}, which corresponds to the 0-dimensional homology.}

\item{p}{An integer value specifying the p-norm to use for the Wasserstein
distance. Defaults to \code{2L}, which corresponds to the Euclidean distance. If
\code{p} is set to \code{Inf}, then the Bottleneck distance is used.}

\item{ncores}{An integer value specifying the number of cores to use when
computing the pairwise distance matrix between all combined persistence
diagrams. Defaults to \code{1L}, which means that the computation is done
sequentially.}

\item{B}{An integer value specifying the number of permutations to use for
the permutation hypothesis test. Defaults to \code{1000L}.}

\item{npc}{A string specifying the non-parametric combination method to use.
Choices are either \code{"tippett"} (default) or \code{"fisher"}. The former
corresponds to the Tippet's method, while the latter corresponds to
Fisher's method.}

\item{verbose}{A boolean value indicating whether to print some information
about the progress of the computation. Defaults to \code{FALSE}.}
}
\value{
A p-value from the two-sample test where the null hypothesis is that
the two samples come from the same distribution.
}
\description{
This function performs a two-sample test for persistence homology data using
the theory of permutation hypothesis testing. The input data can take on
various forms:
\itemize{
\item A persistence set, which is a collection of persistence diagrams.
\item A distance matrix, which is a pairwise distance matrix between persistence
diagrams.
\item One of the PH vectorizations available in the
\href{https://cran.r-project.org/package=TDAvec}{{TDAvec}} package.
}
}
\examples{
two_sample_test(trefoils1[1:5], trefoils2[1:5], B = 100L)
two_sample_test(trefoils1[1:5], archspirals[1:5], B = 100L)
}
