% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inphr-package.R
\docType{package}
\name{inphr-package}
\alias{inphr}
\alias{inphr-package}
\title{inphr: Inference for Persistence Homology Data in R}
\description{
A set of functions for performing null hypothesis testing on samples of persistence diagrams using the theory of permutations. Currently, only two-sample testing is implemented. Inputs can be either samples of persistence diagrams themselves or vectorizations. In the former case, they are embedded in a metric space using either the Bottleneck or Wasserstein distance. In the former case, persistence data becomes functional data and inference is performed using tools available in the 'fdatest' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tdaverse/inphr}
  \item \url{https://tdaverse.github.io/inphr/}
  \item Report bugs at \url{https://github.com/tdaverse/inphr/issues}
}

}
\author{
\strong{Maintainer}: Aymeric Stamm \email{aymeric.stamm@cnrs.fr} (\href{https://orcid.org/0000-0002-8725-3654}{ORCID})

}
\keyword{internal}
